Fix in printf

+++ tag/printf.h
@@ -256,7 +256,7 @@ namespace tag
 		{
 			static void print(std::ostream& o, const std::string& fmt, int fpos, const T_list<C,D>& l)
 			{
-				unsigned int ppos;
+				size_t ppos;
 
 				while(1)
 				{
